add 06-1, change 94, 98

+++ 98-purge-users.sh
@@ -6,8 +6,12 @@
 
 read -p "Press ENTER to continue"
 
-for i in {1..12}
+for i in {2..32}
 do
+	if ! [ -d "/home/catbot-$i" ]; then
+		echo "No catbot $i";
+		continue;
+	fi
 	echo "Deleting user catbot-$i"
 	sudo userdel -r catbot-$i
 	sudo groupdel catbot-$i
